backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
dj-database-url==3.0.1
python-dotenv==1.2.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
